Use VotingKeyChanged events to determine voter set.
Instead of `InitializeChange`, which may or may not lead to an actual
change of the current voter set later, use `VotingKeyChanged` to track
the set.

Also, check the contract addresses, so only events from the correct
instance of the contracts are handled.

The `-v` option now displays the full lists of participating and
abstaining voters for each ballot.

And the Readme recommends the release binaries, so users don't need to
have Cargo installed.